REFACTOR :: 폴더구조
