+++ apps/user/src/components/form/Tables/index.ts
@@ -1,0 +1,8 @@
+export { default as AttendanceCalculator } from './AttendanceCalculator/AttendanceCalculator';
+export { default as CertificateCalculator } from './CertificateCalculator/CertificateCalculator';
+export { default as GradeCalculator } from './GradeCalculator/GradeCalculator';
+export { default as GradeCalculatorHeader } from './GradeCalculatorHeader/GradeCalculatorHeader';
+export { default as GradeCalculatorItem } from './GradeCalculatorItem/GradeCalculatorItem';
+export { default as GradePreview } from './GradePreview/GradePreview';
+export { default as NewGradeCalculatorItem } from './NewGradeCalculatorItem/NewGradeCalculatorItem';
+export { default as VolunteerCalculator } from './VolunteerCalculator/VolunteerCalculator';
