+++ apps/user/src/components/form/index.ts
@@ -1,9 +1,6 @@
-export { default as 보호자정보 } from './apps/보호자정보/보호자정보';
-export { default as 지원자정보 } from './apps/지원자정보/지원자정보';
-export { default as 출신학교및학력 } from './apps/출신학교및학력/출신학교및학력';
-export { default as 성적입력 } from './apps/성적입력/성적입력';
-export { default as 전형선택 } from './apps/전형선택/전형선택';
-export { default as 자기소개서 } from './apps/자기소개서/자기소개서';
-
-export { default as ProgressBar } from './common/ProgressBar/ProgressBar';
-export { default as FormController } from './common/FormController/FormController';
+export { default as FindAddressModal } from './FindAddressModal/FindAddressModal';
+export { default as FindSchoolModal } from './FindSchoolModal/FindSchoolModal';
+export { default as ProfileUpload } from './ProfileUpload/ProfileUpload';
+export { default as ProgressBar } from './ProgressBar/ProgressBar';
+export { default as FormController } from './FormController/FormController';
+export * from './Tables';
