+++ packages/maru-ui/index.ts
@@ -1,5 +1,5 @@
 export { default as Button } from './components/Button/Button';
-export { default as DropdownMenu } from './components/Dropdown/Dropdown';
+export { default as Dropdown } from './components/Dropdown/Dropdown';
 export { default as Column } from './components/Flex/Column';
 export { default as Row } from './components/Flex/Row';
 export { default as ButtonInput } from './components/Input/ButtonInput';
